refactor: rename ResourceFileConstants to DataFileConstants; move scores.txt to data/
ResourceFileConstants held only HighscoreFilePath after audio constants were
extracted to AudioResourceConstants. It is renamed to DataFileConstants to
accurately reflect that it holds filesystem data file paths, not classpath
resource paths.

HighscoreFilePath updated from 'scores.txt' to 'data/scores.txt':
  - maze/scores.txt         -> maze/data/scores.txt
  - test fixture also moved -> maze-javafx-backend/src/test/resources/data/scores.txt

writeScore() in ActionScreenController now calls mkdirs() on the parent
directory before writing, so the data/ directory is created automatically
on first run if it does not exist.

All callers updated: ActionScreenController, HighScoreController, GdxGameScreen.
