+++ .gitignore
@@ -5,7 +5,6 @@ Thumbs.db
 
 #General:
 *.class
-scores.txt
 
 # Maven build output:
 **/target/
