+++ maze-javafx-backend/src/test/resources/data/scores.txt
